feat: reusable responsive modal shell, rewire
